+++ lib/types.rs
@@ -1,7 +1,15 @@
 //! Useful types used across multiple Falcon modules.
 
+use analysis::calling_convention::{CallingConvention, CallingConventionType};
 use il;
 use translator;
+
+/// A boolean type with an unknown value
+pub enum PartialBoolean {
+    True,
+    False,
+    Unknown
+}
 
 /// The underlying endianness of this memory model.
 #[derive(Clone, Debug, Deserialize, Eq, PartialEq, Serialize)]
@@ -20,6 +28,7 @@ pub enum Architecture {
 
 
 impl Architecture {
+    /// Get the endianness of the given architecture.
     pub fn endian(&self) -> Endian {
         match *self {
             Architecture::X86 |
@@ -28,6 +37,7 @@ impl Architecture {
         }
     }
 
+    /// Get the translator/lifter for this architecture.
     pub fn translator(&self) -> Box<translator::Translator> {
         match *self {
             Architecture::X86 => Box::new(translator::x86::X86::new()),
@@ -36,11 +46,23 @@ impl Architecture {
         }
     }
 
+    /// Get the default calling convention for this Architecture
+    pub fn calling_convention(&self) -> CallingConvention {
+        match *self {
+            Architecture::Mips |
+            Architecture::Mipsel =>
+                CallingConvention::new(CallingConventionType::MipsSystemV),
+            Architecture::X86 =>
+                CallingConvention::new(CallingConventionType::Cdecl),
+        }
+    }
+
+    /// Get the stack pointer for this architecture
     pub fn stack_pointer(&self) -> il::Scalar {
         match *self {
-            Architecture::X86 => il::scalar("esp", 32),
             Architecture::Mips |
-            Architecture::Mipsel => il::scalar("$sp", 32)
+            Architecture::Mipsel => il::scalar("$sp", 32),
+            Architecture::X86 => il::scalar("esp", 32)
         }
     }
 }
@@ -51,6 +73,8 @@ fn test_x86() {
     let arch = Architecture::X86;
     assert_eq!(arch.endian(), Endian::Little);
     assert_eq!(arch.stack_pointer(), il::scalar("esp", 32));
+    assert_eq!(*arch.calling_convention()
+                    .return_register(), il::scalar("eax", 32));
 }
 
 #[test]
@@ -58,6 +82,8 @@ fn test_mips() {
     let arch = Architecture::Mips;
     assert_eq!(arch.endian(), Endian::Big);
     assert_eq!(arch.stack_pointer(), il::scalar("$sp", 32));
+    assert_eq!(*arch.calling_convention()
+                    .return_register(), il::scalar("$v0", 32));
 }
 
 #[test]
@@ -65,4 +91,6 @@ fn test_mipsel() {
     let arch = Architecture::Mipsel;
     assert_eq!(arch.endian(), Endian::Little);
     assert_eq!(arch.stack_pointer(), il::scalar("$sp", 32));
+    assert_eq!(*arch.calling_convention()
+                    .return_register(), il::scalar("$v0", 32));
 }